Add LibrarySectioner struct
- This will allow it to have some state.
